backend/requirements.txt
Django==3.1.6
djangorestframework==3.12.2
psycopg2-binary==2.8.6
python-decouple==3.4
django-paranoid==0.8
django-allauth==0.44.0
